Meta2: Reduce the exposure of bean generation functions
